Review comments Move modules
